Fix an issue where the card reader service would stop and hang
This issue is due to the fact that we clear the
`$rootScope.serialTimeout` when changing routes, and starting a new
polling loop. When the operation of starting a new poll loop happens
inbetween a read() operation has started but not finished, the new loop
will also start a read simultaniously. This causes us to have two read
operations, where one will finish successfully and the other will wait
for new data. Essentially leaving a "gap" in the reads for the new loop.

This commit fixes this issue by using a simple lock to determine when
the serial device has started a read which is not complete. Any other
loops will then break out of the reading loop until the device is
available and ready and the lock is available.